refactor(cli): move debug collection into lib
